Extract email tests to test_mail_subcribe.py
