+++ eventex/subscriptions/tests/test_view_subscribe.py
@@ -55,32 +55,6 @@ class SubscribePostValid(TestCase):
 	def test_send_subscribe_email(self):
 		self.assertEqual(1, len(mail.outbox))
 
-	def test_subscription_email_subject(self):
-		email = mail.outbox[0]
-		expect = 'Confirmação de inscrição'
-
-		self.assertEqual(expect, email.subject)
-
-	def test_subscription_email_from(self):
-		email = mail.outbox[0]
-		expect = 'contato@eventex.com.br'
-
-		self.assertEqual(expect, email.from_email)
-
-	def test_subscription_email_to(self):
-		email = mail.outbox[0]
-		expect = ['contato@eventex.com.br', 'henrique@bastos.net']
-
-		self.assertEqual(expect, email.to)
-
-	def test_subscription_email_body(self):
-		email = mail.outbox[0]
-
-		self.assertIn('Henrique Bastos', email.body)
-		self.assertIn('12345678901', email.body)
-		self.assertIn('henrique@bastos.net', email.body)
-		self.assertIn('21-99618-6180', email.body)
-
 
 class SubscribePostInvalid(TestCase):
 	def setUp(self):
